Move calculateBuffor() method to separate class
