+++ client/src/main/java/com/battleships/gamewindow/board/Fleet.java
@@ -1,8 +1,6 @@
 package com.battleships.gamewindow.board;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 
 public class Fleet {
