+++ client/src/main/java/com/battleships/gamewindow/board/PlayerBoard.java
@@ -4,7 +4,6 @@ import com.battleships.Coordinate;
 import com.battleships.gamewindow.board.fieldStates.BoardField;
 import com.battleships.gamewindow.board.fieldStates.FieldState;
 import com.battleships.gamewindow.services.RandomFleetPlacement;
-import javafx.scene.layout.GridPane;
 
 import java.util.HashSet;
 import java.util.List;
@@ -37,12 +36,6 @@ public class PlayerBoard extends Board {
         }
     }
 
-    private void addBoardFieldToView(GridPane gridPaneBoard, Coordinate mastCoord, BoardField boardField) {
-        int fieldColumn = mastCoord.getColumn();
-        int fieldRow = mastCoord.getRow();
-        gridPaneBoard.add(boardField, fieldColumn, fieldRow);
-    }
-
     public void markButtonsAsHit(Coordinate[] coordinates) {
         BoardField boardField = board.get(coordinates[0]);
         boardField.hit();
@@ -52,25 +45,8 @@ public class PlayerBoard extends Board {
         ship.getMasts().forEach(boardField -> board.put(boardField.getCoordinate(), boardField));
     }
 
-    public Set<BoardField> calculateBuffer(Ship ship) {
-        Set<BoardField> buffer = new HashSet<>();
-        Set<BoardField> masts = ship.getMasts();
-
-        masts.forEach(boardField -> {
-
-            for (int rowIncrement = -1; rowIncrement < 2; rowIncrement++) {
-                for (int colIncrement = -1; colIncrement < 2; colIncrement++) {
-                    BoardField incrementedBordField = getBoardField(boardField.getCoordinate(), rowIncrement, colIncrement);
-                    if (incrementedBordField != null && incrementedBordField.isSea())
-                        buffer.add(incrementedBordField);
-                }
-            }
-
-        });
-        return buffer;
+    public BoardField getField(Coordinate coordinate) {
+        return board.get(coordinate);
     }
 
-    private BoardField getBoardField(Coordinate coords, int plusRow, int plusCol) {
-        return board.get(Coordinate.fromIntCoords(coords.getRow() + plusRow, coords.getColumn() + plusCol));
-    }
 }
